pyproject.toml
[tool.poetry]
name = "kafka-streaming-demo"
version = "0.1.0"
description = "Kafka streaming demo including a producer, a consumer and windowing"
authors = ["zsofiak96"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
confluent-kafka = "^2.5.3"
pyspark = "^3.5.2"
requests = "^2.32.3"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
schedule = "^1.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
pre-commit = "^3.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
